[WM] use base Docker image and run rsyslog from supervisor
I was experiencing strange errors with many of the docker.io-built
containers, but this works locally and on docker1 when I push my image
to quay.io, pull it, and run it there.

So if this fails from the wercker build, we should try using quay.io for
the base image instead of docker.io.